- cast to unsigned int for all calls to srand

+++ source/glest_game/main/intro.cpp
@@ -236,20 +236,20 @@ Intro::Intro(Program *program):
 
 			//unsigned int seed = time(NULL);
 			Chrono seed(true);
-			srand(seed.getCurTicks());
+			srand((unsigned int)seed.getCurTicks());
 			int failedLookups=0;
 			std::map<int,bool> usedIndex;
 			for(;modelList.size() < models.size();) {
 				int index = rand() % models.size();
 				if(usedIndex.find(index) != usedIndex.end()) {
 					failedLookups++;
-					srand(seed.getCurTicks() / failedLookups);
+					srand((unsigned int)seed.getCurTicks() / failedLookups);
 					continue;
 				}
 				//printf("picIndex = %d list count = %d\n",picIndex,coreData.getMiscTextureList().size());
 				modelList.push_back(models[index]);
 				usedIndex[index]=true;
-				srand(seed.getCurTicks() / modelList.size());
+				srand((unsigned int)seed.getCurTicks() / modelList.size());
 			}
 			models = modelList;
 		}
@@ -426,14 +426,14 @@ Intro::Intro(Program *program):
 		if(showIntroPicsRandom == true) {
 			//unsigned int seed = time(NULL);
 			Chrono seed(true);
-			srand(seed.getCurTicks());
+			srand((unsigned int)seed.getCurTicks());
 			int failedLookups=0;
 			std::map<int,bool> usedIndex;
 			for(;intoTexList.size() < showIntroPics;) {
 				int picIndex = rand() % coreData.getMiscTextureList().size();
 				if(usedIndex.find(picIndex) != usedIndex.end()) {
 					failedLookups++;
-					srand(seed.getCurTicks() / failedLookups);
+					srand((unsigned int)seed.getCurTicks() / failedLookups);
 
 					if(failedLookups > 10000) {
 						for(unsigned int i = 0;
@@ -451,7 +451,7 @@ Intro::Intro(Program *program):
 				//printf("picIndex = %d list count = %d\n",picIndex,coreData.getMiscTextureList().size());
 				intoTexList.push_back(coreData.getMiscTextureList()[picIndex]);
 				usedIndex[picIndex]=true;
-				srand(seed.getCurTicks() / intoTexList.size());
+				srand((unsigned int)seed.getCurTicks() / intoTexList.size());
 			}
 		}
 		else {
