+++ source/glest_game/menu/menu_state_masterserver.cpp
@@ -250,9 +250,8 @@ MenuStateMasterserver::MenuStateMasterserver(Program *program, MainMenu *mainMen
 	GraphicComponent::applyAllCustomProperties(containerName);
 
     char szIRCNick[80]="";
-    //srand(time(NULL));
 	Chrono seed(true);
-	srand(seed.getCurTicks());
+	srand((unsigned int)seed.getCurTicks());
 
     int randomNickId = rand() % 999;
     string netPlayerName=Config::getInstance().getString("NetPlayerName",Socket::getHostName().c_str());
